Keyword upfront

+++ src/core/utils/color-utils.js
@@ -24,34 +24,56 @@ export function getSolidColors(color) {
 
 
 
+// const sp = new SuperWorkers.MainThread({
+//   url: 'js/image-worker.js',
+//   maxWorkers: 5,
+// });
+
+const imgQueue = [];
+let workers = 0;
+const MAX_WORKERS = 3;
 const processCanvas = new OffscreenCanvas(1200, 1200);
 const processCtx = processCanvas.getContext("2d");
 export function processImage(src, cb) {
-  
-  const worker = new Worker('/image-worker.js');
-  worker.addEventListener('message', (e) => {
-    if(e.data.bitmap) {
-      const bitmap = e.data.bitmap;
-      processCtx.drawImage(bitmap, 0, 0);
-      const imgd = processCtx.getImageData(0, 0, bitmap.width, bitmap.height);
-      
-      const data = {
-        // src,
-        buf8Buffer: imgd.data.buffer, 
-        imgSize: {
-          width: bitmap.width, 
-          height: bitmap.height,
-        },
+  let callback = cb;
+  if(workers >= MAX_WORKERS) {
+    imgQueue.push({src, cb});
+  } else {
+    workers++;
+    const worker = new Worker('js/image-worker.js');
+    worker.addEventListener('message', (e) => {
+      if(e.data.bitmap) {
+        const bitmap = e.data.bitmap;
+        processCtx.drawImage(bitmap, 0, 0);
+        const imgd = processCtx.getImageData(0, 0, bitmap.width, bitmap.height);
+        
+        const data = {
+          // src,
+          buf8Buffer: imgd.data.buffer, 
+          imgSize: {
+            width: bitmap.width, 
+            height: bitmap.height,
+          },
+        }
+        worker.postMessage(data, [imgd.data.buffer]);
       }
-      worker.postMessage(data, [imgd.data.buffer]);
-    }
-    else if(e.data.palette) {
-      const { palette } = e.data;
-      cb({palette});
-      worker.terminate();
-    }
-  })
-  worker.postMessage({src});
+      else if(e.data.palette) {
+        const { palette } = e.data;
+        callback({palette});
+
+        if(imgQueue.length) {
+          const next = imgQueue.shift();
+          callback = next.cb;
+          worker.postMessage({src: next.src});
+        } else {
+          workers--;
+          worker.terminate();
+        }
+      }
+    })
+    worker.postMessage({src});
+  }
+}
 
   
   // const imgEl = new Image();
@@ -91,4 +113,3 @@ export function processImage(src, cb) {
   //     }
   //     worker.postMessage(data, [imgd.data.buffer]);
   //   })
-}
